requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
pypdf
openai
python-dotenv
torch
transformers